% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobRead.R
\name{get_var_names}
\alias{get_var_names}
\title{Function for returning the variable names within a specified reanalysis dataset.}
\usage{
get_var_names(dataset)
}
\arguments{
\item{dataset}{One of ERA5, JRA55 or MERRA2 specifying which reanalysis to use.}
}
\value{
A vector of GlobSim climate varible names specific to a given reanalysis dataset.
}
\description{
This function extracts the climate variable names present within a scaled GlobSim nc dataset.
}
