% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobRead.R
\name{station_names}
\alias{station_names}
\title{Simple function to print station names form within specified dataset, should be same for all datasets}
\usage{
station_names(dataset)
}
\arguments{
\item{dataset}{One of ERA5, JRA55 or MERRA2 specifying which reanalysis to use, in this case results should be the same.}
}
\value{
A vector of GlobSim station 'site' names specific to a given reanalysis dataset.
}
\description{
This function extracts the station 'site' names present within a scaled GlobSim nc dataset.
}
